refactor(timeline): split timeline state and small-state events out of mod.rs
Stage 2 of the room_screen split, verbatim code motion: the
TimelineUpdate protocol, the per-room TimelineUiState cache and its
save/restore types move to state.rs; small-state event grouping, the
summary phrasing, and the SmallStateEventContent populate pass move to
small_state.rs. The public types (RoomScreenProps,
RoomScreenTooltipActions, TimelineUpdate, clear_timeline_states) are
re-exported from mod.rs, so every external path is unchanged.
mod.rs: 13,831 -> 12,659 lines.
