+++ src/home/room_screen/mod.rs
@@ -61,11 +61,16 @@ use super::{ContextMenuOpenGesture, event_reaction_list::ReactionData, invite_mo
 mod bot_admin;
 mod bot_message;
 mod octos_actions;
+mod small_state;
+mod state;
 
 use bot_admin::*;
 use bot_message::*;
 use octos_actions::*;
+use small_state::*;
+use state::*;
 pub(crate) use bot_admin::is_known_or_likely_bot;
+pub use state::{RoomScreenProps, RoomScreenTooltipActions, TimelineUpdate, clear_timeline_states};
 
 /// The maximum number of timeline items to search through
 /// when looking for a particular event.
@@ -157,7 +162,6 @@ const BOT_BADGE_HORIZONTAL_PADDING: f64 = 6.0;
 const BOT_BADGE_BORDER_RADIUS: f64 = 3.0;
 const BOT_BADGE_TEXT_FONT_SIZE: f64 = 8.5;
 const BOT_BADGE_TEXT_TOP_DROP: f64 = -0.08;
-const MIN_SMALL_STATE_EVENTS_TO_COLLAPSE: usize = 2;
 
 const fn centered_top_margin(outer_top_margin: f64, outer_height: f64, inner_height: f64) -> f64 {
     outer_top_margin + ((outer_height - inner_height) * 0.5)
@@ -9874,772 +9878,6 @@ impl RoomScreenRef {
     }
 }
 
-/// Immutable RoomScreen states passed via Scope props
-/// from a RoomScreen widget to its child widgets for event/draw handlers.
-pub struct RoomScreenProps {
-    pub room_screen_widget_uid: WidgetUid,
-    pub room_name_id: RoomNameId,
-    pub timeline_kind: TimelineKind,
-    pub room_members: Option<Arc<Vec<RoomMember>>>,
-    pub is_encrypted: Option<bool>,
-    pub is_direct_room: bool,
-    pub room_bot_user_ids: Vec<OwnedUserId>,
-    pub room_members_sync_pending: bool,
-    /// Pre-computed sort order for room members (for mention search optimization).
-    pub room_members_sort: Option<Arc<crate::room::member_search::PrecomputedMemberSort>>,
-    pub room_avatar_url: Option<OwnedMxcUri>,
-    pub app_service_enabled: bool,
-    pub app_service_room_bound: bool,
-    pub has_persisted_management_binding: bool,
-    pub bound_bot_user_id: Option<OwnedUserId>,
-    pub resolved_parent_bot_user_id: Option<OwnedUserId>,
-    pub persisted_bound_bot_user_ids: Vec<OwnedUserId>,
-    pub known_bot_user_ids: Vec<OwnedUserId>,
-    /// Whether the current user has permission to invite users to this room.
-    /// Gates the `/invitebot` slash command.
-    pub can_invite: bool,
-    /// Invites this client has sent that are still awaiting acceptance —
-    /// consumed by the `/invitebot` picker (at open time) to keep its
-    /// candidate filtering idempotent during the invite round-trip.
-    pub pending_invited_users: Vec<OwnedUserId>,
-}
-
-
-/// Actions for the room screen's tooltip.
-#[derive(Clone, Debug, Default)]
-pub enum RoomScreenTooltipActions {
-    /// Mouse over event when the mouse is over the read receipt.
-    HoverInReadReceipt {
-        /// The rect of the moused over widget
-        widget_rect: Rect,
-        /// Includes the list of users who have seen this event
-        read_receipts: indexmap::IndexMap<matrix_sdk::ruma::OwnedUserId, Receipt>,
-    },
-    /// Mouse over event when the mouse is over the reaction button.
-    HoverInReactionButton {
-        /// The rectangle (bounds) of the hovered-over widget.
-        widget_rect: Rect,
-        /// Includes the list of users who have reacted to the emoji.
-        reaction_data: ReactionData,
-    },
-    /// Mouse out event and clear tooltip.
-    HoverOut,
-    #[default]
-    None,
-}
-
-/// A message that is sent from a background async task to a room's timeline view
-/// for the purpose of update the Timeline UI contents or metadata.
-pub enum TimelineUpdate {
-    /// The very first update a given room's timeline receives.
-    FirstUpdate {
-        /// The initial list of timeline items (events) for a room.
-        initial_items: Vector<Arc<TimelineItem>>,
-    },
-    /// The content of a room's timeline was updated in the background.
-    NewItems {
-        /// The entire list of timeline items (events) for a room.
-        new_items: Vector<Arc<TimelineItem>>,
-        /// The range of indices in the `items` list that have been changed in this update
-        /// and thus must be removed from any caches of drawn items in the timeline.
-        /// Any items outside of this range are assumed to be unchanged and need not be redrawn.
-        changed_indices: Range<usize>,
-        /// An optimization that informs the UI whether the changes to the timeline
-        /// resulted in new items being *appended to the end* of the timeline.
-        is_append: bool,
-        /// Whether to clear the entire cache of drawn items in the timeline.
-        /// This supersedes `index_of_first_change` and is used when the entire timeline is being redrawn.
-        clear_cache: bool,
-    },
-    /// The updated number of unread messages in the room.
-    NewUnreadMessagesCount(UnreadMessageCount),
-    /// The target event ID was found at the given `index` in the timeline items vector.
-    ///
-    /// This means that the RoomScreen widget can scroll the timeline up to this event,
-    /// and the background `timeline_subscriber_handler` async task can stop looking for this event.
-    TargetEventFound {
-        target_event_id: OwnedEventId,
-        index: usize,
-    },
-    /// A notice that the background task doing pagination for this room is currently running
-    /// a pagination request in the given direction, and is waiting for that request to complete.
-    PaginationRunning(PaginationDirection),
-    /// An error occurred while paginating the timeline for this room.
-    PaginationError {
-        error: timeline::Error,
-        direction: PaginationDirection,
-    },
-    /// A notice that the background task doing pagination for this room has become idle,
-    /// meaning that it has completed its recent pagination request(s).
-    PaginationIdle {
-        /// If `true`, the start of the timeline has been reached, meaning that
-        /// there is no need to send further pagination requests.
-        fully_paginated: bool,
-        direction: PaginationDirection,
-    },
-    /// A notice that event details have been fetched from the server,
-    /// including a `result` that indicates whether the request was successful.
-    EventDetailsFetched {
-        event_id: OwnedEventId,
-        result: Result<(), matrix_sdk_ui::timeline::Error>,
-    },
-    /// A notice that fresh thread-summary details were fetched for a thread root.
-    ThreadSummaryDetailsFetched {
-        thread_root_event_id: OwnedEventId,
-        timeline_item_index: usize,
-        num_replies: u32,
-        latest_reply_preview_text: Option<String>,
-    },
-    /// The result of a request to edit a message in this timeline.
-    MessageEdited {
-        timeline_event_item_id: TimelineEventItemId,
-        result: Result<(), matrix_sdk_ui::timeline::Error>,
-    },
-    /// A notice that the room's members have been fetched from the server,
-    /// though the success or failure of the request is not yet known until the client
-    /// requests the member info via a timeline event's `sender_profile()` method.
-    RoomMembersSynced,
-    /// A notice that the room's full member list has been fetched from the server,
-    /// includes a complete list of room members that can be shared across components.
-    /// This is different from RoomMembersSynced which only indicates members were fetched
-    /// but doesn't provide the actual data.
-    RoomMembersListFetched {
-        members: Vec<RoomMember>,
-    },
-    /// A notice with an option of Media Request Parameters that one or more requested media items (images, videos, etc.)
-    /// that should be displayed in this timeline have now been fetched and are available.
-    MediaFetched(MediaRequestParameters),
-    /// A notice that one or more members of a this room are currently typing.
-    TypingUsers {
-        /// The list of display names of users who are currently typing in this room.
-        users: Vec<String>,
-    },
-    /// The result of a pin/unpin request ([`MatrixRequest::PinEvent`]).
-    PinResult {
-        event_id: OwnedEventId,
-        result: Result<bool, matrix_sdk::Error>,
-        pin: bool,
-    },
-    /// An update containing the set of pinned events in this room.
-    PinnedEvents(Vec<OwnedEventId>),
-    /// An update containing the currently logged-in user's power levels for this room.
-    UserPowerLevels(UserPowerLevels),
-    /// An update to the currently logged-in user's own read receipt for this room.
-    OwnUserReadReceipt(Receipt),
-    /// A notice that the given room has been tombstoned (closed)
-    /// and replaced by the given successor room.
-    Tombstoned(SuccessorRoomDetails),
-    /// A notice that link preview data for a URL has been fetched and is now available.
-    LinkPreviewFetched,
-    /// User confirmed a file upload via the file upload modal.
-    FileUploadConfirmed(crate::shared::file_upload_modal::FileData),
-    /// Progress update for an ongoing file upload.
-    FileUploadUpdate {
-        current: u64,
-        total: u64,
-    },
-    /// The abort handle for an in-progress file upload.
-    FileUploadAbortHandle(tokio::task::AbortHandle),
-    /// An error occurred during file upload.
-    FileUploadError {
-        error: String,
-        file_data: crate::shared::file_upload_modal::FileData,
-        retryable: bool,
-    },
-    /// File upload completed successfully.
-    FileUploadComplete,
-    /// A file/media attachment download has completed (or failed) for this timeline.
-    ///
-    /// The `Result` indicates whether the save operation succeeded or failed.
-    /// This does not immediately clear the pending-download entry so the UI can
-    /// briefly show success/failure state.
-    AttachmentDownloadFinished(OwnedMxcUri, Result<(), String>),
-    /// Remove the given pending-download entry and return to idle button state.
-    AttachmentDownloadReset(OwnedMxcUri),
-}
-
-fn enqueue_timeline_update(
-    pending_updates: &mut VecDeque<TimelineUpdate>,
-    update: TimelineUpdate,
-) {
-    match update {
-        TimelineUpdate::NewItems {
-            new_items,
-            changed_indices,
-            clear_cache,
-            is_append,
-        } => {
-            if let Some(TimelineUpdate::NewItems {
-                changed_indices: previous_changed_indices,
-                clear_cache: previous_clear_cache,
-                is_append: previous_is_append,
-                ..
-            }) = pending_updates.back()
-            {
-                let changed_indices = previous_changed_indices.start.min(changed_indices.start)
-                    ..previous_changed_indices.end.max(changed_indices.end);
-                let clear_cache = *previous_clear_cache || clear_cache;
-                let is_append = *previous_is_append || is_append;
-                *pending_updates.back_mut().expect("checked above") = TimelineUpdate::NewItems {
-                    new_items,
-                    changed_indices,
-                    clear_cache,
-                    is_append,
-                };
-            } else {
-                pending_updates.push_back(TimelineUpdate::NewItems {
-                    new_items,
-                    changed_indices,
-                    clear_cache,
-                    is_append,
-                });
-            }
-        }
-        update => {
-            pending_updates.push_back(update);
-        }
-    }
-}
-
-thread_local! {
-    /// The global set of all timeline states, one entry per room.
-    ///
-    /// This is only useful when accessed from the main UI thread.
-    static TIMELINE_STATES: RefCell<HashMap<TimelineKind, TimelineUiState>> = 
-        RefCell::new(HashMap::new());
-}
-
-/// The UI-side state of a single room's timeline, which is only accessed/updated by the UI thread.
-///
-/// This struct should only include states that need to be persisted for a given room
-/// across multiple `Hide`/`Show` cycles of that room's timeline within a RoomScreen.
-/// If a state is more temporary and shouldn't be persisted when the timeline is hidden,
-/// then it should be stored in the RoomScreen widget itself, not in this struct.
-struct TimelineUiState {
-    /// Info determining whether this is a main room timeline is a thread-focused timeline.
-    kind: TimelineKind,
-
-    /// The power levels of the currently logged-in user in this room.
-    user_power: UserPowerLevels,
-
-    /// The list of room members for this room.
-    room_members: Option<Arc<Vec<RoomMember>>>,
-
-    /// Pre-computed sort order for room members (for efficient mention search).
-    room_members_sort: Option<Arc<crate::room::member_search::PrecomputedMemberSort>>,
-
-    /// Whether the initial room-member sync is still in progress for this room.
-    room_members_sync_pending: bool,
-
-    /// Whether we're waiting for a refreshed local member snapshot after sync completion.
-    awaiting_post_sync_member_refresh: bool,
-
-    /// Whether this room's timeline has been fully paginated, which means
-    /// that the oldest (first) event in the timeline is locally synced and available.
-    /// When `true`, further backwards pagination requests will not be sent.
-    ///
-    /// This must be reset to `false` whenever the timeline is fully cleared.
-    fully_paginated: bool,
-
-    /// Whether a backwards pagination request has already been submitted
-    /// and is still in flight.
-    backwards_pagination_in_flight: bool,
-
-    /// The list of items (events) in this room's timeline that our client currently knows about.
-    items: Vector<Arc<TimelineItem>>,
-
-    /// The set of first-event IDs for small-state event groups that are expanded.
-    ///
-    /// By default, groups are collapsed unless their first event ID appears in this set.
-    expanded_small_state_group_event_ids: HashSet<OwnedEventId>,
-
-    /// Derived lookup used by the draw path to fold small state-event groups.
-    ///
-    /// This is invalidated only when timeline items or group expansion state
-    /// changes, instead of rebuilding an O(timeline length) index every frame.
-    small_state_event_group_index: Option<SmallStateEventGroupIndex>,
-
-    /// Event IDs of long bot messages the user chose to unfold.
-    ///
-    /// Long agent replies are folded to a short preview by default; an ID lands
-    /// here only after the user taps "show more". Kept on the timeline state
-    /// (not the recycled item widget) so the choice survives PortalList
-    /// virtualization, exactly like `expanded_small_state_group_event_ids`.
-    expanded_bot_body_event_ids: HashSet<OwnedEventId>,
-
-    /// The range of items (indices in the above `items` list) whose event **contents** have been drawn
-    /// since the last update and thus do not need to be re-populated on future draw events.
-    ///
-    /// This range is partially cleared on each background update (see below) to ensure that
-    /// items modified during the update are properly redrawn. Thus, it is a conservative
-    /// "cache tracker" that may not include all items that have already been drawn,
-    /// but that's okay because big updates that clear out large parts of the rangeset
-    /// only occur during back pagination, which is both rare and slow in and of itself.
-    /// During typical usage, new events are appended to the end of the timeline,
-    /// meaning that the range of already-drawn items doesn't need to be cleared.
-    ///
-    /// Upon a background update, only item indices greater than or equal to the
-    /// `index_of_first_change` are removed from this set.
-    content_drawn_since_last_update: RangeSet<usize>,
-
-    /// Same as `content_drawn_since_last_update`, but for the event **profiles** (avatar, username).
-    profile_drawn_since_last_update: RangeSet<usize>,
-
-    /// The channel receiver for timeline updates for this room.
-    ///
-    /// Here we use a synchronous (non-async) channel because the receiver runs
-    /// in a sync context and the sender runs in an async context,
-    /// which is okay because a sender on an unbounded channel never needs to block.
-    update_receiver: crossbeam_channel::Receiver<TimelineUpdate>,
-
-    /// Updates already pulled from the channel but deferred to a later UI pass.
-    pending_updates: VecDeque<TimelineUpdate>,
-
-    /// The sender for timeline requests from a RoomScreen showing this room
-    /// to the background async task that handles this room's timeline updates.
-    request_sender: TimelineRequestSender,
-
-    /// Coordinates pagination requests sent by the room list and this timeline UI.
-    pagination_status: Arc<TimelinePaginationStatus>,
-
-    /// The cache of media items (images, videos, etc.) that appear in this timeline.
-    ///
-    /// Currently this excludes avatars, as those are shared across multiple rooms.
-    media_cache: MediaCache,
-
-    /// Cache for link preview data indexed by URL to avoid redundant network requests.
-    link_preview_cache: LinkPreviewCache,
-    /// Cached fetched thread-summary details, keyed by thread-root event ID.
-    fetched_thread_summaries: HashMap<OwnedEventId, FetchedThreadSummary>,
-    /// Set of thread roots currently being fetched to avoid duplicate in-flight requests.
-    pending_thread_summary_fetches: HashSet<OwnedEventId>,
-
-    /// The states relevant to the UI display of this timeline that are saved upon
-    /// a `Hide` action and restored upon a `Show` action.
-    saved_state: SavedState,
-
-    /// The state of the message highlight animation.
-    ///
-    /// We need to run the animation once the scrolling, triggered by the click of of a
-    /// a reply preview, ends. so we keep a small state for it.
-    /// By default, it starts in Off.
-    /// Once the scrolling is started, the state becomes Pending.
-    /// If the animation was triggered, the state goes back to Off.
-    message_highlight_animation_state: MessageHighlightAnimationState,
-
-    /// Active streaming animations, keyed by event ID.
-    /// Stores the typewriter animation state for messages being streamed by bots.
-    streaming_messages: HashMap<OwnedEventId, super::streaming_animation::StreamingAnimState>,
-
-    /// The index of the timeline item that was most recently scrolled up past it.
-    /// This is used to detect when the user has scrolled up past the second visible item (index 1)
-    /// upwards to the first visible item (index 0), which is the top of the timeline,
-    /// at which point we submit a backwards pagination request to fetch more events.
-    last_scrolled_index: usize,
-
-    /// The index of the first item shown in the timeline's PortalList from *before* the last "jump".
-    ///
-    /// This index is saved before the timeline undergoes any jumps, e.g.,
-    /// receiving new items, major scroll changes, or other timeline view jumps.
-    prev_first_index: Option<usize>,
-
-    /// Whether the user has scrolled past their latest read marker.
-    ///
-    /// This is used to determine whether we should send a fully-read receipt
-    /// after the user scrolls past their "read marker", i.e., their latest fully-read receipt.
-    /// Its value is determined by comparing the fully-read event's timestamp with the
-    /// first and last timestamp of displayed events in the timeline.
-    /// When scrolling down, if the value is true, we send a fully-read receipt
-    /// for the last visible event in the timeline.
-    ///
-    /// When new message come in, this value is reset to `false`.
-    scrolled_past_read_marker: bool,
-    latest_own_user_receipt: Option<Receipt>,
-
-    /// If `Some`, this room has been tombstoned and the details of its successor room
-    /// are contained within. If `None`, the room has not been tombstoned.
-    tombstone_info: Option<SuccessorRoomDetails>,
-    /// Media attachments currently being downloaded in this timeline.
-    pending_downloads: Vec<PendingDownload>,
-}
-
-#[derive(Default, Debug)]
-enum MessageHighlightAnimationState {
-    Pending { item_id: usize },
-    #[default]
-    Off,
-}
-
-/// States that are necessary to save in order to maintain a consistent UI display for a timeline.
-///
-/// These are saved when navigating away from a timeline (upon `Hide`)
-/// and restored when navigating back to a timeline (upon `Show`).
-#[derive(Default)]
-struct SavedState {
-    /// The index of the first item in the timeline's PortalList that is currently visible,
-    /// and the scroll offset from the top of the list's viewport to the beginning of that item.
-    /// If this is `None`, then the timeline has not yet been scrolled by the user
-    /// and the portal list will be set to "tail" (track) the bottom of the list.
-    first_index_and_scroll: Option<(usize, f64)>,
-    /// The state of all UI elements in the `RoomInputBar`.
-    room_input_bar_state: RoomInputBarState,
-}
-
-#[derive(Clone, Debug)]
-struct SmallStateEventGroup {
-    start: usize,
-    end: usize,
-    count: usize,
-    first_event_id: OwnedEventId,
-    collapsed: bool,
-}
-
-#[derive(Default)]
-struct SmallStateEventGroupIndex {
-    by_start: HashMap<usize, SmallStateEventGroup>,
-    collapsed_hidden_indices: RangeSet<usize>,
-    summary_by_start: HashMap<usize, String>,
-}
-
-#[derive(Default)]
-struct SmallStateSummaryStats {
-    joined_users: Vec<String>,
-    left_users: Vec<String>,
-    profile_picture_changes: HashMap<String, usize>,
-    display_name_changes: HashMap<String, usize>,
-    other_changes: usize,
-}
-
-fn timeline_item_is_small_state_event(
-    timeline_item: &TimelineItem,
-    timeline_kind: &TimelineKind,
-) -> bool {
-    let TimelineItemKind::Event(event_tl_item) = timeline_item.kind() else {
-        return false;
-    };
-    match event_tl_item.content() {
-        TimelineItemContent::MsgLike(msg_like_content) => {
-            if timeline_kind.thread_root_event_id().is_none()
-                && msg_like_content.thread_root.is_some()
-            {
-                return false;
-            }
-            matches!(
-                msg_like_content.kind,
-                MsgLikeKind::Poll(_)
-                | MsgLikeKind::UnableToDecrypt(_)
-                | MsgLikeKind::LiveLocation(_)
-                | MsgLikeKind::Other(_)
-            )
-        }
-        TimelineItemContent::MembershipChange(_)
-        | TimelineItemContent::ProfileChange(_)
-        | TimelineItemContent::OtherState(_) => true,
-        _ => false,
-    }
-}
-
-fn compute_small_state_event_groups(
-    items: &Vector<Arc<TimelineItem>>,
-    timeline_kind: &TimelineKind,
-    expanded_group_event_ids: &HashSet<OwnedEventId>,
-) -> Vec<SmallStateEventGroup> {
-    let mut groups = Vec::new();
-    let mut idx = 0usize;
-    while idx < items.len() {
-        let is_small = items
-            .get(idx)
-            .is_some_and(|item| timeline_item_is_small_state_event(item, timeline_kind));
-        if !is_small {
-            idx += 1;
-            continue;
-        }
-
-        let start = idx;
-        idx += 1;
-        while idx < items.len()
-            && items
-                .get(idx)
-                .is_some_and(|item| timeline_item_is_small_state_event(item, timeline_kind))
-        {
-            idx += 1;
-        }
-        let end = idx;
-        let count = end.saturating_sub(start);
-        if count < MIN_SMALL_STATE_EVENTS_TO_COLLAPSE {
-            continue;
-        }
-
-        let Some(first_event_id) = items
-            .get(start)
-            .and_then(|item| item.as_event())
-            .and_then(|event| event.event_id())
-            .map(ToOwned::to_owned)
-        else {
-            continue;
-        };
-
-        groups.push(SmallStateEventGroup {
-            start,
-            end,
-            count,
-            collapsed: !expanded_group_event_ids.contains(&first_event_id),
-            first_event_id,
-        });
-    }
-    groups
-}
-
-fn index_small_state_event_groups(
-    groups: impl IntoIterator<Item = SmallStateEventGroup>,
-) -> SmallStateEventGroupIndex {
-    let mut index = SmallStateEventGroupIndex::default();
-    for group in groups {
-        if group.collapsed {
-            index
-                .collapsed_hidden_indices
-                .insert(group.start + 1 .. group.end);
-        }
-        index.by_start.insert(group.start, group);
-    }
-    index
-}
-
-fn build_small_state_event_group_index(
-    items: &Vector<Arc<TimelineItem>>,
-    timeline_kind: &TimelineKind,
-    expanded_group_event_ids: &HashSet<OwnedEventId>,
-    app_language: AppLanguage,
-) -> SmallStateEventGroupIndex {
-    let mut index = index_small_state_event_groups(compute_small_state_event_groups(
-        items,
-        timeline_kind,
-        expanded_group_event_ids,
-    ));
-    index.summary_by_start = index
-        .by_start
-        .iter()
-        .map(|(&start, group)| {
-            (
-                start,
-                format_small_state_group_summary_text(app_language, items, group),
-            )
-        })
-        .collect();
-    index
-}
-
-fn prune_expanded_small_state_group_ids(
-    items: &Vector<Arc<TimelineItem>>,
-    timeline_kind: &TimelineKind,
-    expanded_group_event_ids: &mut HashSet<OwnedEventId>,
-) {
-    let empty_expanded_ids: HashSet<OwnedEventId> = HashSet::new();
-    let valid_group_ids: HashSet<OwnedEventId> = compute_small_state_event_groups(
-        items,
-        timeline_kind,
-        &empty_expanded_ids,
-    )
-    .into_iter()
-    .map(|group| group.first_event_id)
-    .collect();
-    expanded_group_event_ids.retain(|event_id| valid_group_ids.contains(event_id));
-}
-
-fn summarize_sender_name(event_tl_item: &EventTimelineItem) -> String {
-    if let TimelineDetails::Ready(profile) = event_tl_item.sender_profile()
-        && let Some(name) = profile.display_name.as_ref()
-        && !name.is_empty()
-    {
-        return name.clone();
-    }
-
-    let raw = event_tl_item.sender().as_str();
-    let without_at = raw.strip_prefix('@').unwrap_or(raw);
-    without_at.split(':').next().unwrap_or(without_at).to_string()
-}
-
-fn push_unique_name(names: &mut Vec<String>, name: String) {
-    if !names.iter().any(|n| n == &name) {
-        names.push(name);
-    }
-}
-
-fn collect_small_state_summary_stats(
-    items: &Vector<Arc<TimelineItem>>,
-    group: &SmallStateEventGroup,
-) -> SmallStateSummaryStats {
-    let mut stats = SmallStateSummaryStats::default();
-    for idx in group.start .. group.end {
-        let Some(item) = items.get(idx) else { continue };
-        let TimelineItemKind::Event(event_tl_item) = item.kind() else { continue };
-        let sender_name = summarize_sender_name(event_tl_item);
-        match event_tl_item.content() {
-            TimelineItemContent::MembershipChange(change) => match change.change() {
-                Some(MembershipChange::Joined)
-                | Some(MembershipChange::InvitationAccepted) => {
-                    push_unique_name(&mut stats.joined_users, sender_name);
-                }
-                Some(MembershipChange::Left)
-                | Some(MembershipChange::KnockRetracted)
-                | Some(MembershipChange::InvitationRejected) => {
-                    push_unique_name(&mut stats.left_users, sender_name);
-                }
-                Some(MembershipChange::NotImplemented)
-                | Some(MembershipChange::None)
-                | Some(MembershipChange::Error)
-                | None => {}
-                _ => {
-                    stats.other_changes += 1;
-                }
-            },
-            TimelineItemContent::ProfileChange(change) => {
-                let mut did_count = false;
-                if change.avatar_url_change().is_some() {
-                    *stats.profile_picture_changes.entry(sender_name.clone()).or_insert(0) += 1;
-                    did_count = true;
-                }
-                if change.displayname_change().is_some() {
-                    *stats.display_name_changes.entry(sender_name).or_insert(0) += 1;
-                    did_count = true;
-                }
-                if !did_count {
-                    stats.other_changes += 1;
-                }
-            }
-            TimelineItemContent::OtherState(_)
-            | TimelineItemContent::MsgLike(_) => {
-                stats.other_changes += 1;
-            }
-            _ => {}
-        }
-    }
-    stats
-}
-
-fn format_people_phrase(
-    app_language: AppLanguage,
-    names: &[String],
-    one_suffix_en: &str,
-    plural_suffix_en: &str,
-    one_suffix_zh: &str,
-    plural_suffix_zh: &str,
-) -> Option<String> {
-    if names.is_empty() {
-        return None;
-    }
-    Some(match app_language {
-        AppLanguage::ChineseSimplified => match names.len() {
-            1 => format!("{}{}", names[0], one_suffix_zh),
-            2 => format!("{}、{}{}", names[0], names[1], plural_suffix_zh),
-            n => format!("{} 等 {} 人{}", names[0], n, plural_suffix_zh),
-        },
-        AppLanguage::English => match names.len() {
-            1 => format!("{}{}", names[0], one_suffix_en),
-            2 => format!("{} and one other{}", names[0], plural_suffix_en),
-            n => format!("{} and {} others{}", names[0], n - 1, plural_suffix_en),
-        },
-    })
-}
-
-fn format_top_user_counter_phrase(
-    app_language: AppLanguage,
-    counts: &HashMap<String, usize>,
-    one_en: &str,
-    many_en: &str,
-    one_zh: &str,
-    many_zh: &str,
-) -> Option<String> {
-    if counts.is_empty() {
-        return None;
-    }
-    let mut entries: Vec<(&String, &usize)> = counts.iter().collect();
-    entries.sort_by(|(name_a, count_a), (name_b, count_b)| {
-        count_b.cmp(count_a).then_with(|| name_a.cmp(name_b))
-    });
-    let (name, count) = entries[0];
-    Some(match app_language {
-        AppLanguage::ChineseSimplified => {
-            if *count > 1 {
-                format!("{name}{many_zh}", many_zh = many_zh.replace("{count}", &count.to_string()))
-            } else {
-                format!("{name}{one_zh}")
-            }
-        }
-        AppLanguage::English => {
-            if *count > 1 {
-                format!("{name}{many_en}", many_en = many_en.replace("{count}", &count.to_string()))
-            } else {
-                format!("{name}{one_en}")
-            }
-        }
-    })
-}
-
-fn format_small_state_group_summary_text(
-    app_language: AppLanguage,
-    items: &Vector<Arc<TimelineItem>>,
-    group: &SmallStateEventGroup,
-) -> String {
-    let stats = collect_small_state_summary_stats(items, group);
-    let mut parts = Vec::new();
-
-    if let Some(joined) = format_people_phrase(
-        app_language,
-        &stats.joined_users,
-        " joined",
-        " joined",
-        " 加入了房间",
-        " 加入了房间",
-    ) {
-        parts.push(joined);
-    }
-    if let Some(left) = format_people_phrase(
-        app_language,
-        &stats.left_users,
-        " left",
-        " left",
-        " 离开了房间",
-        " 离开了房间",
-    ) {
-        parts.push(left);
-    }
-    if let Some(profile_pic) = format_top_user_counter_phrase(
-        app_language,
-        &stats.profile_picture_changes,
-        " changed their profile picture",
-        " changed their profile picture {count} times",
-        " 更换了头像",
-        " 更换了头像 {count} 次",
-    ) {
-        parts.push(profile_pic);
-    }
-    if let Some(display_name) = format_top_user_counter_phrase(
-        app_language,
-        &stats.display_name_changes,
-        " changed their display name",
-        " changed their display name {count} times",
-        " 修改了昵称",
-        " 修改了昵称 {count} 次",
-    ) {
-        parts.push(display_name);
-    }
-    if stats.other_changes > 0 {
-        parts.push(match app_language {
-            AppLanguage::ChineseSimplified => format!("另有 {} 条其他状态变更", stats.other_changes),
-            AppLanguage::English => format!("{} other state changes", stats.other_changes),
-        });
-    }
-
-    if parts.is_empty() {
-        return match app_language {
-            AppLanguage::ChineseSimplified => format!("{} 条状态事件", group.count),
-            AppLanguage::English => format!("{} state events", group.count),
-        };
-    }
-    parts.join(", ")
-}
-
 /// Returns info about the item in the list of `new_items` that matches the event ID
 /// of a visible item in the given `curr_items` list.
 ///
@@ -12447,299 +11685,6 @@ pub fn populate_preview_of_timeline_item(
 }
 
 
-/// A trait for abstracting over the different types of timeline events
-/// that can be displayed in a `SmallStateEvent` widget.
-trait SmallStateEventContent {
-    /// Populates the *content* (not the profile) of the given `item` with data from
-    /// the given `event_tl_item` and `self` (the specific type of event content).
-    ///
-    /// ## Arguments
-    /// * `item`: a `SmallStateEvent` widget that has already been added to
-    ///   the given `PortalList` at the given `item_id`.
-    ///   This function may either modify that item or completely replace it
-    ///   with a different widget if needed.
-    /// * `item_drawn_status`: the old (prior) drawn status of the item.
-    /// * `new_drawn_status`: the new drawn status of the item, which may have already
-    ///   been updated to reflect the item's profile having been drawn right before this function.
-    ///
-    /// ## Return
-    /// Returns a tuple of the drawn `item` and its `new_drawn_status`.
-    fn populate_item_content(
-        &self,
-        cx: &mut Cx,
-        list: &mut PortalList,
-        item_id: usize,
-        item: WidgetRef,
-        event_tl_item: &EventTimelineItem,
-        username: &str,
-        item_drawn_status: ItemDrawnStatus,
-        new_drawn_status: ItemDrawnStatus,
-    ) -> (WidgetRef, ItemDrawnStatus);
-}
-
-// For unable to decrypt messages.
-impl SmallStateEventContent for EncryptedMessage {
-    fn populate_item_content(
-        &self,
-        cx: &mut Cx,
-        _list: &mut PortalList,
-        _item_id: usize,
-        item: WidgetRef,
-        _event_tl_item: &EventTimelineItem,
-        username: &str,
-        _item_drawn_status: ItemDrawnStatus,
-        mut new_drawn_status: ItemDrawnStatus,
-    ) -> (WidgetRef, ItemDrawnStatus) {
-        item.label(cx, ids!(event_row.content)).set_text(
-            cx,
-            &text_preview_of_encrypted_message(self).format_with(username, false),
-        );
-        new_drawn_status.content_drawn = true;
-        (item, new_drawn_status)
-    }
-}
-
-// For other message-like content (custom message-like events).
-impl SmallStateEventContent for LiveLocationState {
-    fn populate_item_content(
-        &self,
-        cx: &mut Cx,
-        _list: &mut PortalList,
-        _item_id: usize,
-        item: WidgetRef,
-        _event_tl_item: &EventTimelineItem,
-        username: &str,
-        _item_drawn_status: ItemDrawnStatus,
-        mut new_drawn_status: ItemDrawnStatus,
-    ) -> (WidgetRef, ItemDrawnStatus) {
-        item.label(cx, ids!(event_row.content)).set_text(
-            cx,
-            &format!("{username} shared a live location."),
-        );
-        new_drawn_status.content_drawn = true;
-        (item, new_drawn_status)
-    }
-}
-
-impl SmallStateEventContent for OtherMessageLike {
-    fn populate_item_content(
-        &self,
-        cx: &mut Cx,
-        _list: &mut PortalList,
-        _item_id: usize,
-        item: WidgetRef,
-        _event_tl_item: &EventTimelineItem,
-        username: &str,
-        _item_drawn_status: ItemDrawnStatus,
-        mut new_drawn_status: ItemDrawnStatus,
-    ) -> (WidgetRef, ItemDrawnStatus) {
-        item.label(cx, ids!(event_row.content)).set_text(
-            cx,
-            &text_preview_of_other_message_like(self).format_with(username, false),
-        );
-        new_drawn_status.content_drawn = true;
-        (item, new_drawn_status)
-    }
-}
-
-// TODO: once we properly display polls, we should remove this,
-//       because Polls shouldn't be displayed using the SmallStateEvent widget.
-impl SmallStateEventContent for PollState {
-    fn populate_item_content(
-        &self,
-        cx: &mut Cx,
-        _list: &mut PortalList,
-        _item_id: usize,
-        item: WidgetRef,
-        _event_tl_item: &EventTimelineItem,
-        _username: &str,
-        _item_drawn_status: ItemDrawnStatus,
-        mut new_drawn_status: ItemDrawnStatus,
-    ) -> (WidgetRef, ItemDrawnStatus) {
-        item.label(cx, ids!(event_row.content)).set_text(
-            cx,
-            self.fallback_text().unwrap_or_else(|| self.results().question).as_str(),
-        );
-        new_drawn_status.content_drawn = true;
-        (item, new_drawn_status)
-    }
-}
-
-impl SmallStateEventContent for timeline::OtherState {
-    fn populate_item_content(
-        &self,
-        cx: &mut Cx,
-        list: &mut PortalList,
-        item_id: usize,
-        item: WidgetRef,
-        _event_tl_item: &EventTimelineItem,
-        username: &str,
-        _item_drawn_status: ItemDrawnStatus,
-        mut new_drawn_status: ItemDrawnStatus,
-    ) -> (WidgetRef, ItemDrawnStatus) {
-        let item = if let Some(text_preview) = text_preview_of_other_state(self, false) {
-            item.label(cx, ids!(event_row.content))
-                .set_text(cx, &text_preview.format_with(username, false));
-            new_drawn_status.content_drawn = true;
-            item
-        } else {
-            let item = list.item(cx, item_id, id!(Empty));
-            new_drawn_status = ItemDrawnStatus::new();
-            item
-        };
-        (item, new_drawn_status)
-    }
-}
-
-impl SmallStateEventContent for MemberProfileChange {
-    fn populate_item_content(
-        &self,
-        cx: &mut Cx,
-        _list: &mut PortalList,
-        _item_id: usize,
-        item: WidgetRef,
-        _event_tl_item: &EventTimelineItem,
-        username: &str,
-        _item_drawn_status: ItemDrawnStatus,
-        mut new_drawn_status: ItemDrawnStatus,
-    ) -> (WidgetRef, ItemDrawnStatus) {
-        item.label(cx, ids!(event_row.content)).set_text(
-            cx,
-            &text_preview_of_member_profile_change(self, username, false)
-                .format_with(username, false),
-        );
-        new_drawn_status.content_drawn = true;
-        (item, new_drawn_status)
-    }
-}
-
-impl SmallStateEventContent for RoomMembershipChange {
-    fn populate_item_content(
-        &self,
-        cx: &mut Cx,
-        list: &mut PortalList,
-        item_id: usize,
-        item: WidgetRef,
-        _event_tl_item: &EventTimelineItem,
-        username: &str,
-        _item_drawn_status: ItemDrawnStatus,
-        mut new_drawn_status: ItemDrawnStatus,
-    ) -> (WidgetRef, ItemDrawnStatus) {
-        let Some(preview) = text_preview_of_room_membership_change(self, false) else {
-            // Don't actually display anything for nonexistent/unimportant membership changes.
-            return (
-                list.item(cx, item_id, id!(Empty)),
-                ItemDrawnStatus::new(),
-            );
-        };
-
-        item.label(cx, ids!(event_row.content))
-            .set_text(cx, &preview.format_with(username, false));
-
-        // The invite_user_button is only used for "Knocked" membership change events.
-        item.button(cx, ids!(event_row.invite_user_button)).set_visible(
-            cx,
-            matches!(self.change(), Some(MembershipChange::Knocked)),
-        );
-
-        new_drawn_status.content_drawn = true;
-        (item, new_drawn_status)
-    }
-}
-
-/// Creates, populates, and adds a SmallStateEvent liveview widget to the given `PortalList`
-/// with the given `item_id`.
-///
-/// The content of the returned widget is populated with data from the
-/// given room membership change and its parent `EventTimelineItem`.
-fn populate_small_state_event(
-    cx: &mut Cx,
-    list: &mut PortalList,
-    item_id: usize,
-    timeline_kind: &TimelineKind,
-    app_language: AppLanguage,
-    event_tl_item: &EventTimelineItem,
-    event_content: &impl SmallStateEventContent,
-    item_drawn_status: ItemDrawnStatus,
-    group_header_summary_text: Option<&str>,
-    group_toggle_button_text: Option<&str>,
-) -> (WidgetRef, ItemDrawnStatus) {
-    let mut new_drawn_status = item_drawn_status;
-    let (item, existed) = list.item_with_existed(cx, item_id, id!(SmallStateEvent));
-    // The content of a small state event view may depend on the profile info,
-    // so we can only mark the content as drawn after the profile has been fully drawn and cached.
-    let skip_redrawing_profile = existed && item_drawn_status.profile_drawn;
-    let skip_redrawing_content = skip_redrawing_profile && item_drawn_status.content_drawn;
-    populate_read_receipts(&item, cx, timeline_kind, event_tl_item);
-    if skip_redrawing_content {
-        return (item, new_drawn_status);
-    }
-
-    // If the profile has been drawn, we can just quickly grab the user's display name
-    // instead of having to call `set_avatar_and_get_username` again.
-    let username_opt = skip_redrawing_profile
-        .then(|| get_profile_display_name(event_tl_item))
-        .flatten();
-
-    let username = username_opt.unwrap_or_else(|| {
-        // As a fallback, call `set_avatar_and_get_username` to get the user's display name.
-        let avatar_ref = item.avatar(cx, ids!(event_row.avatar));
-
-        let (username, profile_drawn) = avatar_ref.set_avatar_and_get_username(
-            cx,
-            timeline_kind,
-            event_tl_item.sender(),
-            Some(event_tl_item.sender_profile()),
-            event_tl_item.event_id(),
-            true,
-        );
-        // Draw the timestamp as part of the profile.
-        if let Some(dt) = unix_time_millis_to_datetime(event_tl_item.timestamp()) {
-            item.timestamp(cx, ids!(event_row.left_container.timestamp)).set_date_time(cx, dt);
-        }
-        new_drawn_status.profile_drawn = profile_drawn;
-        username
-    });
-
-    // Proceed to draw the actual event content.
-    let (item, new_drawn_status) = event_content.populate_item_content(
-        cx,
-        list,
-        item_id,
-        item,
-        event_tl_item,
-        &username,
-        item_drawn_status,
-        new_drawn_status,
-    );
-
-    item.button(cx, ids!(event_row.invite_user_button))
-        .set_text(cx, tr_key(app_language, "room_screen.small_state.invite_to_room"));
-    item.view(cx, ids!(group_header))
-        .set_visible(cx, group_toggle_button_text.is_some());
-    item.label(cx, ids!(group_header.group_summary_label))
-        .set_visible(cx, group_header_summary_text.is_some());
-    if let Some(summary_text) = group_header_summary_text {
-        item.label(cx, ids!(group_header.group_summary_label))
-            .set_text(cx, summary_text);
-    }
-    if let Some(button_text) = group_toggle_button_text {
-        item.button(cx, ids!(group_header.state_group_toggle_button)).set_text(cx, button_text);
-    }
-
-    (item, new_drawn_status)
-}
-
-
-/// Returns the display name of the sender of the given `event_tl_item`, if available.
-fn get_profile_display_name(event_tl_item: &EventTimelineItem) -> Option<String> {
-    if let TimelineDetails::Ready(profile) = event_tl_item.sender_profile() {
-        profile.display_name.clone()
-    } else {
-        None
-    }
-}
-
 
 /// Actions related to invites within a room.
 ///
@@ -13387,87 +12332,9 @@ impl MessageRef {
     }
 }
 
-/// Clears all UI-related timeline states for all known rooms.
-///
-/// This function requires passing in a reference to `Cx`,
-/// which isn't used, but acts as a guarantee that this function
-/// must only be called by the main UI thread. 
-pub fn clear_timeline_states(_cx: &mut Cx) {
-    // Clear timeline states cache
-    TIMELINE_STATES.with_borrow_mut(|states| {
-        states.clear();
-    });
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn adjacent_timeline_snapshots_coalesce_without_crossing_control_updates() {
-        let mut pending = VecDeque::new();
-        enqueue_timeline_update(
-            &mut pending,
-            TimelineUpdate::NewItems {
-                new_items: Vector::new(),
-                changed_indices: 4..8,
-                clear_cache: false,
-                is_append: true,
-            },
-        );
-        match pending.back().unwrap() {
-            TimelineUpdate::NewItems {
-                changed_indices,
-                clear_cache,
-                is_append,
-                ..
-            } => {
-                assert_eq!(changed_indices, &(4..8));
-                assert!(!*clear_cache);
-                assert!(*is_append);
-            }
-            _ => panic!("expected timeline snapshot"),
-        }
-
-        enqueue_timeline_update(
-            &mut pending,
-            TimelineUpdate::NewItems {
-                new_items: Vector::new(),
-                changed_indices: 9..12,
-                clear_cache: false,
-                is_append: true,
-            },
-        );
-        assert_eq!(pending.len(), 1);
-        match pending.back().unwrap() {
-            TimelineUpdate::NewItems {
-                changed_indices,
-                clear_cache,
-                is_append,
-                ..
-            } => {
-                assert_eq!(changed_indices, &(4..12));
-                assert!(!*clear_cache);
-                assert!(*is_append);
-            }
-            _ => panic!("expected coalesced timeline snapshot"),
-        }
-
-        enqueue_timeline_update(
-            &mut pending,
-            TimelineUpdate::PaginationRunning(PaginationDirection::Backwards),
-        );
-        enqueue_timeline_update(
-            &mut pending,
-            TimelineUpdate::NewItems {
-                new_items: Vector::new(),
-                changed_indices: 1..2,
-                clear_cache: false,
-                is_append: true,
-            },
-        );
-        assert_eq!(pending.len(), 3);
-    }
 
     #[test]
     fn test_invite_result_belongs_only_to_pending_screen() {
@@ -13478,45 +12345,6 @@ mod tests {
         assert!(invite_result_belongs_to_room_screen(&pending, &invited_user));
         assert!(!invite_result_belongs_to_room_screen(&pending, &other_user));
         assert!(!invite_result_belongs_to_room_screen(&HashSet::new(), &invited_user));
-    }
-
-    #[test]
-    fn small_state_group_index_stores_group_starts_and_collapsed_ranges() {
-        let collapsed_event_id =
-            OwnedEventId::try_from("$collapsed:example.org").unwrap();
-        let expanded_event_id =
-            OwnedEventId::try_from("$expanded:example.org").unwrap();
-        let index = index_small_state_event_groups([
-            SmallStateEventGroup {
-                start: 2,
-                end: 6,
-                count: 4,
-                first_event_id: collapsed_event_id.clone(),
-                collapsed: true,
-            },
-            SmallStateEventGroup {
-                start: 10,
-                end: 12,
-                count: 2,
-                first_event_id: expanded_event_id.clone(),
-                collapsed: false,
-            },
-        ]);
-
-        assert_eq!(index.by_start.len(), 2);
-        assert_eq!(
-            index.by_start.get(&2).map(|group| (&group.first_event_id, group.count)),
-            Some((&collapsed_event_id, 4)),
-        );
-        assert_eq!(
-            index.by_start.get(&10).map(|group| (&group.first_event_id, group.count)),
-            Some((&expanded_event_id, 2)),
-        );
-        assert!(!index.collapsed_hidden_indices.contains(&2));
-        assert!(index.collapsed_hidden_indices.contains(&3));
-        assert!(index.collapsed_hidden_indices.contains(&5));
-        assert!(!index.collapsed_hidden_indices.contains(&6));
-        assert!(!index.collapsed_hidden_indices.contains(&11));
     }
 
     #[test]
